added verifyResponse and logoutResponse to options
verifyResponse is now called to generate the response to the verify
route. Default response is identical to old response. It can now be
overridden in the options.

logoutResponse is now called to generate the response to the logout
route. Default response is identical to old response. It can now be
overridden in the options.